scaffold/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
anthropic>=0.42.0
chromadb>=0.5.0
mcp[cli]>=1.3.0
python-dotenv>=1.0.0
